+++ Tests/PlayerTests/ExtensionsTests.swift
@@ -11,6 +11,16 @@ import Circumspect
 import Nimble
 import XCTest
 
+final class ClampedTests: XCTestCase {
+    func testClamped() {
+        expect((-1).clamped(to: 0...1)).to(equal(0))
+        expect(0.clamped(to: 0...1)).to(equal(0))
+        expect(0.5.clamped(to: 0...1)).to(equal(0.5))
+        expect(1.clamped(to: 0...1)).to(equal(1))
+        expect(2.clamped(to: 0...1)).to(equal(1))
+    }
+}
+
 final class NotificationCenterPublisherTests: XCTestCase {
     func testWithoutObject() throws {
         let notificationCenter = NotificationCenter.default
